+++ src/core/window.c
@@ -19,6 +19,11 @@ int initGLFW(void){
   glfwWindowHint(GLFW_CONTEXT_VERSION_MINOR, GL_VER_MINOR);
 
   glfwWindowHint(GLFW_OPENGL_PROFILE, GLFW_OPENGL_CORE_PROFILE);
+
+#ifdef __APPLE__
+  glfwWindowHint(GLWF_OPENGL_FORWARD_COMPAT, GL_TRUE);
+#endif
+
   return 0;
 }